csrf fix

+++ routes.py
@@ -56,7 +56,8 @@ def topic(id):
 
 @app.route("/chain/<int:id>")
 def chain(id):
-    return render_template("chain.html", username=users.username(), messages=chains.get_related_messages(id), chain_name=chains.get_chain_name(id), topic_name=chains.get_topic_name(id), id=id, topic_id=chains.get_related_topic(id), csfr_token=session["csrf_token"], is_entitled=chains.entitled_to_chain())
+    user_id = users.user_id()
+    return render_template("chain.html", username=users.username(), messages=chains.get_related_messages(id), chain_name=chains.get_chain_name(id), topic_name=chains.get_topic_name(id), id=id, topic_id=chains.get_related_topic(id), csfr_token=session["csrf_token"], is_entitled=chains.entitled_to_chain(id, user_id))
 
 
 @app.route("/message/<int:id>", methods=["GET"])
@@ -64,15 +65,6 @@ def message(id):
     if request.method == "GET":
         return render_template("message.html", username=users.username(), content=messages.get_content(id), sender=messages.get_sender(id), id=id, chain_id=messages.get_related_chain(id))
 
-
-#@app.route("/create_message", methods=["GET", "POST"])
-#def create_message(chain_id):
-#    if request.method == "GET":
-#        return render_template("create_message.html")
-#    if request.method == "POST":
-#        content = request.form["message"]
-#        messages.create_message(users.user_id(), content, chain_id)
-#        return redirect("chain.html", username=users.username(), messages=chains.get_related_messages(chain_id), chain_name=chains.get_chain_name(chain_id), topic_name=chains.get_topic_name(chain_id))
 
 @app.route("/create_chain", methods=["POST"])
 def create_chain():
@@ -94,33 +86,44 @@ def create_topic():
         print(0)
         if users.is_admin():
             topic_name = request.form["topic_name"]
-
             topics.create_topic(topic_name)
             print(1)
             return redirect("/")
 
-@app.route("/delete_message/<int:id>")
-def delete_message(id):
+
+@app.route("/delete_message", methods=["POST"])
+def delete_message():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
+    id = request.form["id"]
     chain_id = messages.get_related_chain(id)
     messages.delete_message(users.user_id(), id)
     return redirect(f"/chain/{chain_id}")
 
-@app.route("/delete_chain/<int:id>")
-def delete_chain(id):
+@app.route("/delete_chain", methods=["POST"])
+def delete_chain():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
+    id = request.form["id"]
     topic_id = chains.get_related_topic(id)
     chains.delete_chain(id, users.user_id())
     return redirect(f"/topic/{topic_id}")
 
-@app.route("/delete_topic/<int:id>")
-def delete_topic(id):
+
+@app.route("/delete_topic", methods=["POST"])
+def delete_topic():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
+    id = request.form["id"]
     topics.delete_topic(id, users.user_id())
     return redirect("/index")
 
 
 @app.route("/edit_message", methods=["POST"])
 def edit_message():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
     id = request.form["id"]
-    chain_id = request.form["chain_id"]
     content = request.form["new_content"]
     chain_id = messages.get_related_chain(id)
     messages.edit_message(id, users.user_id(), content)
@@ -130,6 +133,8 @@ def edit_message():
 
 @app.route("/send_message", methods=["POST"])
 def send_message():
+    if session["csrf_token"] != request.form["csrf_token"]:
+        abort(403)
     id = request.form["id"]
     content = request.form["content"]
     messages.create_message(users.user_id(), content, messages.get_related_chain(id))
